+++ CMakeLists.txt
@@ -6,4 +6,3 @@ set(CMAKE_BUILD_TYPE Release)
 add_subdirectory(Sprocket)
 add_subdirectory(Game)
 add_subdirectory(Anvil)
-add_subdirectory(Runtime)
